Move unprefixed field/method completion to `dot`
